lobby-service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lobby-service"
version = "0.1.0"
description = "Lobby microservice"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Ben Weber" }
]
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic>=2.8,<3.0",
    "pydantic-settings>=2.4,<3.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]",
    "python-dotenv>=1.0,<2.0",
    "alembic",
    "pyjwt",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "httpx>=0.27,<1.0",
    "ruff>=0.6,<1.0",
    "mypy>=1.11,<2.0",
    "requests>=2.32.5",
    "pytest-asyncio",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
asyncio_mode = "auto"

markers = [
    "unit: fast unit tests",
    "integration: API integration tests",
    "system: container/system tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.mypy]
mypy_path = ["src"]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = false